+++ marko/ext/pangu.py
@@ -14,11 +14,13 @@ Example::
 
     from marko import Markdown
 
-    markdown = Markdown(extensions=[PanguExtension])
+    markdown = Markdown(extensions=['pangu'])
     print(markdown(text))
 """
 from __future__ import unicode_literals
 import re
+
+from marko import helpers
 
 CJK_RE = (
     r"\u2e80-\u2eff\u2f00-\u2fdf\u3040-\u309f\u30a0-\u30ff\u3100-\u312f"
@@ -38,5 +40,8 @@ class PanguRendererMixin(object):
         return PANGU_RE.sub('<span class="pangu"></span>', rv)
 
 
-class PanguExtension:
+class Pangu(object):
     renderer_mixins = [PanguRendererMixin]
+
+
+PanguExtension = helpers._Deprecated(Pangu)