+++ tests/spec/__main__.py
@@ -1,3 +1,0 @@
-from .commonmark import main
-
-main()